refactor trpc server/client global error handling
