+++ src/utils/api.ts
@@ -4,7 +4,7 @@
  *
  * We also create a few inference helpers for input and output types.
  */
-import { httpBatchLink, loggerLink } from '@trpc/client';
+import { httpBatchLink, loggerLink, TRPCClientError } from '@trpc/client';
 import { createTRPCNext } from '@trpc/next';
 import superjson from 'superjson';
 
@@ -12,6 +12,7 @@ import type { AppRouter } from '@/server/api/root';
 import type { inferRouterInputs, inferRouterOutputs } from '@trpc/server';
 import type { NextPageContext } from 'next';
 
+import { trpcClientErrorHandler } from '@/server/api/utils';
 import { DEFAULT_STALE_TIME, ONE_DAY_MS } from './constants';
 
 const getBaseUrl = () => {
@@ -79,7 +80,21 @@ export type RouterInputs = inferRouterInputs<AppRouter>;
 export type RouterOutputs = inferRouterOutputs<AppRouter>;
 
 const queryClientConfig = {
-  defaultOptions: { queries: { staleTime: DEFAULT_STALE_TIME } },
+  defaultOptions: {
+    queries: {
+      staleTime: DEFAULT_STALE_TIME,
+      onError: trpcClientErrorHandler,
+      onSuccess: (data: unknown) => {
+        console.log('query client default success', data);
+      },
+    },
+    mutations: {
+      onError: trpcClientErrorHandler,
+      onSuccess: (data: unknown) => {
+        console.log('mutation client default success', data);
+      },
+    },
+  },
 };
 
 const cacheHeaders = {
